히어로 이미지 및 버튼 동작 수정

+++ Models/SessionStats.cs
@@ -24,7 +24,7 @@ namespace DeskWarrior.Models
         public long TotalDamage { get; set; }
 
         [JsonPropertyName("total_gold")]
-        public int TotalGold { get; set; }
+        public long TotalGold { get; set; }
 
         [JsonPropertyName("monsters_killed")]
         public int MonstersKilled { get; set; }
